+++ JsonData/LoginResponse.cs
@@ -6,14 +6,11 @@ namespace TessApi.JsonData {
 
         public string access_token { get; set; }
         public string token_type { get; set; }
-        public int expires_in { get; set; }
+        public long expires_in { get; set; }
         public string refresh_token { get; set; }
-        public int created_at { get; set; }
+        public long created_at { get; set; }
 
         public string TypeAndToken { get { return token_type + " " + access_token; } }
         public DateTime TokenExpiryDate { get { return DateTime.Today + TimeSpan.FromSeconds(expires_in); } }
-
-
     }
-
 }
